move converter file
