+++ SugorokuLibrary/ClientToServer/CloseCreateMessage.cs
@@ -1,4 +1,5 @@
 using Newtonsoft.Json;
+using SugorokuLibrary.ClientToServer.Converters;
 
 namespace SugorokuLibrary.ClientToServer
 {
