+++ SugorokuLibrary/ClientToServer/Converters/CreatePlayerConverter.cs
@@ -2,17 +2,19 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace SugorokuLibrary.ClientToServer
+namespace SugorokuLibrary.ClientToServer.Converters
 {
-	public class GetMatchInfoConverter : JsonConverter
+	public class CreatePlayerConverter : JsonConverter
 	{
 		public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
 		{
-			var createPlayer = (GetMatchInfoMessage) value!;
+			var createPlayer = (CreatePlayerMessage) value!;
 			
 			writer.WriteStartObject();
 			writer.WritePropertyName("methodType");
 			writer.WriteValue(createPlayer.MethodType);
+			writer.WritePropertyName("playerName");
+			writer.WriteValue(createPlayer.PlayerName);
 			writer.WritePropertyName("matchKey");
 			writer.WriteValue(createPlayer.MatchKey);
 			writer.WriteEndObject();
@@ -21,12 +23,12 @@ namespace SugorokuLibrary.ClientToServer
 		public override object ReadJson(JsonReader reader, Type objectType, object? existingValue, JsonSerializer serializer)
 		{
 			JObject jObject = JObject.Load(reader);
-			return new GetMatchInfoMessage((string) jObject["matchKey"]!);
+			return new CreatePlayerMessage((string) jObject["playerName"]!, (string) jObject["matchKey"]!);
 		}
 
 		public override bool CanConvert(Type objectType)
 		{
-			return objectType == typeof(GetMatchInfoMessage);
+			return objectType == typeof(CreatePlayerMessage);
 		}
 	}
 }
