+++ SugorokuLibrary/ClientToServer/Converters/GetMatchInfoConverter.cs
@@ -2,31 +2,31 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace SugorokuLibrary.ClientToServer
+namespace SugorokuLibrary.ClientToServer.Converters
 {
-	public class CloseCreateConverter : JsonConverter
+	public class GetMatchInfoConverter : JsonConverter
 	{
 		public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
 		{
-			var closeCreate = (CloseCreateMessage) value!;
+			var createPlayer = (GetMatchInfoMessage) value!;
 			
 			writer.WriteStartObject();
 			writer.WritePropertyName("methodType");
-			writer.WriteValue(closeCreate.MethodType);
+			writer.WriteValue(createPlayer.MethodType);
 			writer.WritePropertyName("matchKey");
-			writer.WriteValue(closeCreate.MatchKey);
+			writer.WriteValue(createPlayer.MatchKey);
 			writer.WriteEndObject();
 		}
 
 		public override object ReadJson(JsonReader reader, Type objectType, object? existingValue, JsonSerializer serializer)
 		{
 			JObject jObject = JObject.Load(reader);
-			return new CloseCreateMessage((string) jObject["matchKey"]!);
+			return new GetMatchInfoMessage((string) jObject["matchKey"]!);
 		}
 
 		public override bool CanConvert(Type objectType)
 		{
-			return objectType == typeof(CloseCreateMessage);
+			return objectType == typeof(GetMatchInfoMessage);
 		}
 	}
 }
